requirements.txt
streamlit>=1.26.0
Pillow>=9.0.0
numpy>=1.20.0
